literate: extract interface
